tutorialview: the step identifier is unsigned int, not NSUInteger
All three tutorial-type parameters took NSUInteger where their encodings are
I. The decisive evidence was the property they feed: -tutorialStatus encodes
I16@0:8 and -setTutorialStatus: encodes v20@0:8I16, so the step identifier is
a 32-bit unsigned int throughout and the property was wrong as well.

The property never appeared in the sweep's output, and the reason is worth
recording. It is a synthesised accessor with no explicit declaration in the
implementation, so there is nothing for the tool to compare against. The
mismatch was only visible by following the parameters to what they are stored
into. The sweep finds declarations that disagree with the binary; it cannot
find one that was never written.

Six width reports remain, all signedness rather than size.

+++ Project/RBMenuTutorialView.h
@@ -56,7 +56,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @param rootView The clip-root view whose child controls the steps point at.
  * @ghidraAddress 0x139af8
  */
-- (void)showAnimationWithTutorialType:(NSUInteger)tutorialType
+- (void)showAnimationWithTutorialType:(unsigned int)tutorialType
                          withRootView:(nullable UIView *)rootView;
 
 /**
@@ -85,7 +85,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @param rootView The view whose child control the step points at.
  * @ghidraAddress 0x13b8fc
  */
-- (void)startTutorialWithType:(NSUInteger)tutorialType withRootView:(nullable UIView *)rootView;
+- (void)startTutorialWithType:(unsigned int)tutorialType withRootView:(nullable UIView *)rootView;
 
 /**
  * @brief Run the tutorial step of the given type: pick the control it points at, lay out the
@@ -97,7 +97,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @param animation Whether to animate the layout in.
  * @ghidraAddress 0x13ab34
  */
-- (void)startTutorialWithType:(NSUInteger)tutorialType withAnimation:(BOOL)animation;
+- (void)startTutorialWithType:(unsigned int)tutorialType withAnimation:(BOOL)animation;
 
 #pragma mark Properties
 
@@ -285,7 +285,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @ghidraAddress 0x1412d4 (getter)
  * @ghidraAddress 0x1412e4 (setter)
  */
-@property(nonatomic, assign) NSUInteger tutorialStatus;
+@property(nonatomic, assign) unsigned int tutorialStatus;
 
 /**
  * @brief The message content view's width, derived from the iPad idiom.
